Restructures CLI binaries into single crates.
