+++ .gitignore
@@ -1,4 +1,3 @@
 /target
 */target
 errors.err
-Cargo.lock
