+++ src/lib.rs
@@ -1,1 +1,2 @@
 pub mod logger;
+pub mod cli;
